requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
av==11.0.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.2
biopython==1.83
bleach==6.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
constantly==23.10.4
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dashscope==1.20.8
dataclasses-json==0.6.3
datasets==2.16.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.6.1
docx==0.2.4
environs==9.5.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.4.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.32.0
google-crc32c==1.5.0
googleapis-common-protos==1.63.2
googlesearch-python==1.2.4
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.1
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
ifaddr==0.2.0
imbalanced-learn==0.12.0
imblearn==0.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==22.10.0
intel-openmp==2021.4.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.11
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-core==0.2.25
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langchainhub==0.1.21
langsmith==0.1.94
lightning-utilities==0.10.1
lxml==4.9.3
lxml_html_clean==0.2.0
markdown-it-py==3.0.0
markdown2==2.4.11
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
modelscope==1.16.1
modelscope_studio==0.4.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nh3==0.2.14
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.18.1
onnxscript==0.1.0.dev20240313
openai==1.37.1
openpyxl==3.1.2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
oss2==2.18.6
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber
Pillow==10.1.0
platformdirs==4.1.0
posthog==3.5.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pylibsrtp==0.10.0
pymed
pymilvus==2.4.4
pymongo==4.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyspark==3.5.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rank-bm25==0.2.2
readability-lxml==0.8.1
referencing==0.32.0
regex==2023.10.3
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
ruff==0.6.3
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
tinydb==4.8.0
tokenizers==0.15.0
tomlkit==0.12.0
torch
torchaudio
torchmetrics
torchvision
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
transformers==4.38.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
types-python-dateutil==2.8.19.14
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.9
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.3
zhipuai==2.0.1
zipp==3.19.2
zope.interface==6.2
